Update channel_creator.m
Simulations are done with Rayleigh fading where the variable rician_K_factor is taken as 0. It is an unused variable and remainder of a previous coding process. Non-zero rician_K_factor values do not correspond to Rician fading. If user wants to simulate the Rician fading conditions, he/she must alter the code.